Modified the updateDB method to use the new unicode things from TorrentDef.  Also fixed some bugs in the code:
This is what you are doing:

1. Fetch N uninserted torrent's from the database
2. Try to insert those torrents
3. Wait M seconds and goto 1

However, the problem lies in the 'Try'. When an insert fails, for
whatever reason, it is going to show up in the next 'fetch uninserted
torrents'.

So bug #1: you are now in an infinite loop, upgrade will never finish

Also: when N or more torrents failed in this way, you will only fetch
failing torrents each time.  It is possible that there are still
torrents in the database that could be successfully updated but they
will never be fetched.

So bug #2: you are not guaranteed that all updatable torrents will be
           updated

I'm going to fix this by always inserting something into the
InvertedIndex.  I am just going to use the Torrent.name for keywords
when the .torrent file can not be found.